Major refactoring applied to entire package.
I have split the code originally found in `pytest_fauxfactory` into several
different modules, their intentions which I hope will be clear by their names:

* constants: keep any constants used throughout code handlers: functions which
* will be used to handle diffetent marks marks: these are the mark functions
* themselves

The final version of `pytest_fauxfactory` is much smaller, cleaner, and
hopefully easier to maintain and understand.
