+++ Makefile
@@ -40,7 +40,7 @@ package-upload: package
 		twine upload dist/*
 
 test:
-		coverage run --source pytest_fauxfactory -m py.test -v
+		coverage run --source constants,handlers,helpers,marks,pytest_fauxfactory -m py.test -v
 
 test-coverage: test
 		coverage report -m
