+++ constants.py
@@ -1,0 +1,12 @@
+# -*- coding: utf-8 -*-
+"""Constants used by pytest-fauxfactory."""
+STRING_TYPES = (
+    'alpha',
+    'alphanumeric',
+    'cjk',
+    'html',
+    'latin1',
+    'numeric',
+    'utf8',
+    'punctuation',
+)
